Split Client and Pool out into their own files
